test(agent): consolidate runtime-{events,flow,v2}.test into one file
Three test files for AgentRuntime were a historical artifact (the v2
suffix dates from a long-completed rewrite). Merged into a single
runtime.test.ts (~430 LOC, 29 tests) with describe blocks for: registry,
sendMessage validation, run tracking, sendMessage flow (onRunStart,
cancel reason propagation, abort on early-return), and per-session
event subscription.
